refactor(tests): rename tests to uitests
As going forward there will need to be integration and other tests, rename the normal tests dir to uitests
